+++ app/src/main/java/com/song/game/soaring/bean/attribute/SpellsAttack.java
@@ -1,7 +1,5 @@
 package com.song.game.soaring.bean.attribute;
 
-import android.util.Log;
-
 import com.song.game.soaring.bean.character.Character;
 import com.song.game.soaring.bean.condition.Condition;
 
@@ -21,7 +19,7 @@ public class SpellsAttack implements Attribute {
     }
 
     private float getRate(Character character) {
-        int rate = 0;
+        float rate = 0;
         rate += character.talent.spellsAttackIncrease;
         if (character.condition != null && !character.condition.isEmpty()) {
             for (Condition condition : character.condition) {
